Add column type and remove cell options field editor
