--- src/components/CellOptionsEditor/index.ts
@@ -1,1 +1,0 @@
-export * from './CellOptionsEditor';
