+++ src/components/index.ts
@@ -1,4 +1,3 @@
-export * from './CellOptionsEditor';
 export * from './CellRenderer';
 export * from './GroupsEditor';
 export * from './TablePanel';
