+++ src/module.ts
@@ -1,36 +1,18 @@
-import { identityOverrideProcessor, PanelPlugin } from '@grafana/data';
+import { PanelPlugin } from '@grafana/data';
 
-import { CellOptionsEditor, GroupsEditor, TablePanel } from './components';
-import { CellType, PanelOptions } from './types';
+import { GroupsEditor, TablePanel } from './components';
+import { PanelOptions } from './types';
 
 /**
  * Panel Plugin
  */
-export const plugin = new PanelPlugin<PanelOptions>(TablePanel)
-  .useFieldConfig({
-    useCustomConfig: (builder) => {
-      builder.addCustomEditor({
-        id: 'cellOptions',
-        path: 'cellOptions',
-        name: 'Cell Options',
-        category: ['Cell Options'],
-        defaultValue: {
-          type: CellType.AUTO,
-        },
-        editor: CellOptionsEditor,
-        override: CellOptionsEditor,
-        process: identityOverrideProcessor,
-        shouldApply: () => true,
-      });
-    },
-  })
-  .setPanelOptions((builder) => {
-    builder.addCustomEditor({
-      id: 'groups',
-      path: 'groups',
-      name: 'Tree View based on data source query',
-      editor: GroupsEditor,
-    });
-
-    return builder;
+export const plugin = new PanelPlugin<PanelOptions>(TablePanel).useFieldConfig({}).setPanelOptions((builder) => {
+  builder.addCustomEditor({
+    id: 'groups',
+    path: 'groups',
+    name: 'Column Groups',
+    editor: GroupsEditor,
   });
+
+  return builder;
+});
